Add virtual machine implementation
Implement a basic virtual machine (VM) for interpreting bytecode chunks:
- Add vm package with VM struct and core operations
- Implement stack-based execution model with Push, Pop, and Interpret functions
- Add arithmetic operations (add, subtract, multiply, divide, negate)
- Integrate VM with existing chunk and debug systems
- Update chunk package to use uint8 for opcodes instead of custom OpCode type
- Add new instructions to support arithmetic operations in chunk package
- Modify debug package to handle new instruction set

The VM executes compiled chunks by reading bytecode instructions and operating on a stack. It supports tracing execution when debugTraceExecution is enabled.

This change establishes the foundation for a bytecode interpreter and enables executing simple arithmetic expressions.

+++ internal/chunk/chunk_test.go
@@ -20,21 +20,20 @@ func TestNew(t *testing.T) {
 	if c.lines != nil {
 		t.Errorf("Expected nil lines, got %v", c.lines)
 	}
-	if c.constants.Count() != 0 { // Changed from .count to .Count()
+	if c.constants.Count() != 0 {
 		t.Errorf("Expected constants count 0, got %d", c.constants.Count())
 	}
 }
 
 func TestWrite(t *testing.T) {
 	c := New()
-	if err := c.Write(OP_RETURN, 123); err != nil {
+	if err := c.Write(uint8(OP_RETURN), 123); err != nil {
 		t.Fatalf("Write failed: %v", err)
 	}
-
 	if c.count != 1 {
 		t.Errorf("Expected count 1, got %d", c.count)
 	}
-	if c.Code()[0] != OP_RETURN {
+	if c.Code()[0] != uint8(OP_RETURN) {
 		t.Errorf("Expected OP_RETURN, got %d", c.Code()[0])
 	}
 	if c.Lines()[0] != 123 {
@@ -48,7 +47,7 @@ func TestAddConstant(t *testing.T) {
 	if idx != 0 {
 		t.Errorf("Expected index 0, got %d", idx)
 	}
-	if c.constants.Count() != 1 { // Changed from .count to .Count()
+	if c.constants.Count() != 1 {
 		t.Errorf("Expected constants count 1, got %d", c.constants.Count())
 	}
 	if c.constants.Values()[0] != 1.2 {
@@ -59,7 +58,7 @@ func TestAddConstant(t *testing.T) {
 func TestGrow(t *testing.T) {
 	c := New()
 	for i := 0; i < 10; i++ {
-		if err := c.Write(OP_RETURN, 123); err != nil {
+		if err := c.Write(uint8(OP_RETURN), 123); err != nil {
 			t.Fatalf("Write failed: %v", err)
 		}
 	}
@@ -73,12 +72,11 @@ func TestGrow(t *testing.T) {
 
 func TestFree(t *testing.T) {
 	c := New()
-	if err := c.Write(OP_RETURN, 123); err != nil {
+	if err := c.Write(uint8(OP_RETURN), 123); err != nil {
 		t.Fatalf("Write failed: %v", err)
 	}
 	c.AddConstant(value.Value(1.2))
 	c.Free()
-
 	if c.count != 0 {
 		t.Errorf("Expected count 0, got %d", c.count)
 	}
@@ -91,7 +89,7 @@ func TestFree(t *testing.T) {
 	if c.lines != nil {
 		t.Errorf("Expected nil lines, got %v", c.lines)
 	}
-	if c.constants.Count() != 0 { // Changed from .count to .Count()
+	if c.constants.Count() != 0 {
 		t.Errorf("Expected constants count 0, got %d", c.constants.Count())
 	}
 }
